requirements.txt
streamlit==1.38.0
langchain==0.2.17
pypdf==4.3.1
langchain-openai==0.1.23
python-dotenv==1.0.1
faiss-cpu==1.8.0
openai==1.43.0
redis==5.2.0
langchain-community==0.2.19
langchain-redis